# docs/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SALFORD backend API — proposed contract",
    "version": "1.0.0-design",
    "description": "Design artifact only; no backend service is implemented. All paths are relative to /v1. UUIDs are server identifiers, never Figma node ids. Authenticated requests derive the user from the bearer token. See docs/architecture.md for invariants and conditional billing channels."
  },
  "servers": [
    {
      "url": "https://api.example.com/v1",
      "description": "Placeholder; replace before use"
    },
    {
      "url": "http://10.0.2.2:3000/v1",
      "description": "Proposed Android emulator development endpoint; no server is running"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/billing/purchases/verify": {
      "post": {
        "operationId": "verifyPurchase",
        "tags": [
          "Billing"
        ],
        "summary": "Queue verification of a native store purchase",
        "description": "Require verified account. Verify actual provider purchase/product/package/bundle/environment/account binding. Same evidence cannot belong to two users. Restore uses this same endpoint for each store-owned purchase.",
        "x-phase": "core",
        "x-roles": [
          "student",
          "admin"
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Verification"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Same key + body returns original response; changed body returns 409. Minimum retention 24h."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequest"
              }
            }
          }
        }
      }
    },
    "/billing/verifications/{verificationId}": {
      "get": {
        "operationId": "getVerification",
        "tags": [
          "Billing"
        ],
        "summary": "Poll verification result after purchase or restore",
        "description": "Poll verification result after purchase or restore",
        "x-phase": "core",
        "x-roles": [
          "student",
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Verification"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "parameters": [
          {
            "name": "verificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/me/subscriptions/{subscriptionId}/management-session": {
      "post": {
        "operationId": "manageSubscription",
        "tags": [
          "Billing"
        ],
        "summary": "Open provider-managed cancellation or plan change",
        "description": "Apple/Google return allowlisted subscription-management links. Enabled external provider returns a short-lived customer portal URL. Do not promise cancellation until verified provider state arrives.",
        "x-phase": "core",
        "x-roles": [
          "student",
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ManageBilling"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/checkout-sessions": {
      "post": {
        "operationId": "createCheckout",
        "tags": [
          "Billing"
        ],
        "summary": "Create optional external hosted checkout",
        "description": "Disabled unless deployment/distribution policy and merchant availability are approved. Return 403 CHANNEL_NOT_ALLOWED otherwise. Server controls amount, currency, customer and allowlisted return URLs. Never accept PAN/CVV or arbitrary redirect URLs.",
        "x-phase": "conditional",
        "x-roles": [
          "student",
          "admin"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Checkout"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Same key + body returns original response; changed body returns 409. Minimum retention 24h."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "productId"
                ]
              }
            }
          }
        }
      }
    },
    "/billing/checkout-sessions/{checkoutId}": {
      "get": {
        "operationId": "getCheckout",
        "tags": [
          "Billing"
        ],
        "summary": "Poll verified hosted checkout state",
        "description": "Only succeeded after provider verification and subscription transaction. Client redirect/success callback is not proof of payment.",
        "x-phase": "conditional",
        "x-roles": [
          "student",
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Checkout"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/admin/plans/{planId}/products": {
      "post": {
        "operationId": "mapProduct",
        "tags": [
          "Admin"
        ],
        "summary": "Map an existing provider product to a plan",
        "description": "Validate provider product exists and matches environment and interval. Cannot invent a price in client requests. Plan cannot become active until required store product mappings exist.",
        "x-phase": "admin",
        "x-roles": [
          "admin"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWrite"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "pubsubBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google OIDC JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Product": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "enum": [
              "apple",
              "google",
              "stripe"
            ]
          },
          "productId": {
            "type": "string"
          },
          "offerId": {
            "type": "string"
          },
          "interval": {
            "type": "string",
            "enum": [
              "month",
              "year"
            ]
          },
          "displayPrice": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "amountMinor": {
            "type": "integer",
            "minimum": 0,
            "nullable": true
          },
          "environment": {
            "type": "string",
            "enum": [
              "sandbox",
              "production"
            ]
          }
        },
        "required": [
          "id",
          "provider",
          "productId",
          "offerId",
          "interval",
          "displayPrice",
          "currency",
          "amountMinor",
          "environment"
        ]
      },
...
      "ApplePurchase": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "apple"
            ]
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "signedTransaction": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16000,
            "writeOnly": true
          }
        },
        "required": [
          "provider",
          "productId",
          "signedTransaction"
        ]
      },
      "GooglePurchase": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "google"
            ]
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "purchaseToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16000,
            "writeOnly": true
          }
        },
        "required": [
          "provider",
          "productId",
          "purchaseToken"
        ]
      },
      "PurchaseRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ApplePurchase"
          },
          {
            "$ref": "#/components/schemas/GooglePurchase"
          }
        ]
      },
      "Verification": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "rejected"
            ]
          },
          "subscriptionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "failureCode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "subscriptionId",
          "failureCode"
        ]
      },
      "Checkout": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "succeeded",
              "failed",
              "expired"
            ]
          },
          "checkoutUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "subscriptionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "checkoutUrl",
          "expiresAt",
          "subscriptionId"
        ]
      },
      "ManageBilling": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "provider": {
            "type": "string",
            "enum": [
              "apple",
              "google",
              "stripe"
            ]
          }
        },
        "required": [
          "url",
          "expiresAt",
          "provider"
        ]
      },
...
      "ProductWrite": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "apple",
              "google",
              "stripe"
            ]
          },
          "environment": {
            "type": "string",
            "enum": [
              "sandbox",
              "production"
            ]
          },
          "productId": {
            "type": "string",
            "maxLength": 255
          },
          "offerId": {
            "type": "string",
            "maxLength": 255
          },
          "interval": {
            "type": "string",
            "enum": [
              "month",
              "year"
            ]
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "provider",
          "environment",
          "productId",
          "offerId",
          "interval",
          "active"
        ]
      },
...
    },
    "responses": {
      "ApiError": {
        "description": "Structured error; 429 includes Retry-After. Ownership failures use 404; unauthenticated requests use 401.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Only for 429/temporary failures"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}